moved my account page components to a folder as well as product details components
